requirements.txt
django==5.2.1
django-ninja==1.4.1
django-ninja-jwt==5.3.7
structlog==25.3.0
django-ninja-extra>=0.22.9
pydantic>=2.11.0
gunicorn==21.2.0
psycopg2-binary==2.9.9 